+++ types/simulation/config.go
@@ -33,12 +33,6 @@ func (c Config) shallowCopy() Config {
 	return c
 }
 
-// With sets the values of t, seed, and fuzzSeed in a copy of the Config and returns the copy.
-func (c Config) With(tb testing.TB, seed int64, fuzzSeed []byte) Config {
-	tb.Helper()
-	r := c.shallowCopy()
-	r.TB = tb
-	r.Seed = seed
-	r.FuzzSeed = fuzzSeed
-	return r
+	DBBackend   string // custom db backend type
+	BlockMaxGas int64  // custom max gas for block
 }